requirements.txt
transformers>=4.40.0
accelerate>=0.29.0
bitsandbytes>=0.43.0
peft>=0.10.0
sentence-transformers>=2.7.0

llmlingua>=0.2.2
selective-context>=0.1.0

xgboost>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

tiktoken>=0.6.0
nltk>=3.8.1
pyyaml>=6.0.1
tqdm>=4.66.0
requests>=2.31.0

pandas>=2.1.0
matplotlib>=3.8.0
